remove the "News from INRIA" section
The INRIA website is only infrequently updated. I see his prominent
place on the front page as a waste of valuable space that could be
better used. Hopefully the caml.inria.fr news are aggregated on Planet
OCaml already, so they should appear in the "News from the community"
anyway.

Remark: I left the "News from the community" heading instead of
reverting to the old "News", as I think the "from the community" part
lightly underlines that the content appearing below is not endorsed by
ocaml-lang.org as it may have been written by anyone without much
editorial control.