Strengthen the Weberizer bindings against failure of the renderers
I just tried to compile ocaml.org from a place where any access to
a webpage sends me the default HTML of a captive web portal. The build
process tries to access the RSS of the OCaml planet, and fails
horribly when parsing the HTML blurb from the captive portal. The
failure of a single weberizer renderer should not stop the rendering
of the whole website.

The current patches catch those renderer failure, report them on
stderr during the build, but keep rendering the rest of the
website. The generated HTML indicates the binding call that resulted
in the error (eg. "rss http://.../rss20.xml").

This feature *might* benefit from being moved upstream to Weberizer,
but I'm not sure there would be a good catch-all fallback strategy for
unknown renderers.