rename OCamlWeb_Main to Main
The OCamlWeb_ prefix was following the convention sometimes used of
starting module names with the library name. This is unnecessary in
this project because the library is never intended to be installed; it
is only for internal use by the build scripts. Thus, we might as well
use simpler module names.