Update functional_programming.html
Grammar and syntax modifications. Note that the word alright was not modified for the word all right, even though the former is not an "actual word", as using all right would not have the same "ring" to it, and thus would not give justice to the sentence. Also, the link at the end is broken and does not lead anywhere: http://ocaml-tutorial.org/_bin/upload_image_form.cmo?name=boxedarray.png