Launch a separate toplevel for each HTML file.
The toplevel is executed in a separate process which can be started
and stopped at will (for each HTML file that executes a toplevel
sentence ATM).  The advantages are that the camlp4 extensions loaded
for a page do not influence the other pages and that the main "build"
program can be compiled to native code.