Improve body text contrast.
The current choice of slightly gray text on white is very hard to read on some
devices.  This fixes the colors to be black on white, to make large bodies of
text and the tutorials easier on the eyes.

(I made this same change to realworldocaml.org after several requests from
reviewers)