Fix "from source installation method"
There should be at least a link on that page to go to `/releases/`.

I removed “This is
unlikely the method you want to use unless you are a developer of the
OCaml compiler itself or trying to install on some esoteric platforms.”
because there are many cases when one would prefer getting the original
sources.