Update 99problems.md
Proposing a more direct implementation of (exact k list) which also prunes the search and outputs the choices in a more canonical order. The original is a bit hard to follow for me, so I'm not sure if it prunes or not. The original seems to generate a lot of heap garbage to me but maybe OCaml knows how to optimize that out.