Advertise utop more prominently
This proposed  change comes from the discussion in ocaml/ocaml#241 , where @jwatzman points out that `utop` does not have enough visibility in the "Basics" tutorial -- it is mentioned, but maybe it should be emphasized more.

It is hard to be informative yet concise enough: I realize this Basics tutorial is the first approach of OCaml content, so it should be to the point and should not drown the reader in pointless distinctions. Maybe the utop can be made more visible (than the previous mention) but more concise (than this proposed change).

We may need to differentiate better between a page that would be "a tour of the language" (let me show you why you should care, without necessarily having you run code) and "basic setup if you know you want to use". Currently the Basics tutorial tries to do both and it is non-optimal at them both.