replace links to svn.html with github
Links from old release pages, e.g. releases/3.12.1.html, not updated.
Those pages are in some sense archival, and I thought better to leave
them as-is. The svn.html page already mentions that the info there is
out of date, so there is no risk someone will end up following the link
and not realizing GitHub is the new location.