corrected graphics code
The graphics code was not correct and it is not compilable. According to ocaml (v>=4.05) `let ()` must return with a value of type `unit`, but the code provided here was using `read_line ()` which is of type `string -> unit`. So I used `print_endline ""` which is of type `unit` to correct the code.