Change the order of the suggested Ubuntu packages.
It seems that students looking to install OCaml on Ubuntu will find this page, not know what X11 means and proceed to install using the first suggested command. By swapping the two suggested commands, we ensure that people who don't know what they are doing are getting X11 support and the Graphics package.