Use cohttp as HTTP client for planet and CWN
The ocamlnet+gnutls combination is confused by servers that expose
TLS1.3 resulting in errors like `GNUTLS_E_PUSH_ERROR`. This causes many
feeds from planet to fail.

This replaces it with cohttp using openssl as TLS client.
Ocaml-tls is a possible alternative, but when used with cohttp it does
not check server certificates, so it unsuitable for this task.

A drawback is that cohttp does not have redirect logic, so it's
necessary to implement it in here.